Система лайков BJS2-16729 - сделал validation и фикс Advice
