+++ src/main/java/faang/school/postservice/dto/like/LikeDto.java
@@ -1,10 +1,12 @@
 package faang.school.postservice.dto.like;
 
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 @Data
 public class LikeDto {
     private Long id;
+    @NotNull
     private Long userId;
     private Long commentId;
     private Long postId;
